requirements.txt
async-generator==1.10
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.11
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
docutils==0.20
et-xmlfile==1.1.0
exceptiongroup==1.1.1
exchangelib==4.9.0
fire==0.4.0
fonttools==4.39.4
fpdf2==2.7.4
furl==2.1.3
graphviz==0.13.2
h11==0.14.0
htmldocx==0.0.6
hubspot-api-client==4.0.6
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
isodate==0.6.1
java-access-bridge-wrapper==0.9.7
jsonpath-ng==1.5.3
jsonschema==4.17.3
lxml==4.9.2
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.3.3
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.2
orderedmultidict==1.0.1
outcome==1.2.0
packaging==23.1
pdfminer.six==20201018
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
ply==3.11
proxy-tools==0.1.0
psutil==5.9.5
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
pynput-robocorp-fork==4.0.0
pyotp==2.8.0
PyPDF2==1.28.6
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
PySocks==1.7.1
pyspnego==0.9.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-xlib==0.33
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywebview==3.6.3
PyYAML==5.4.1
requests==2.30.0
requests-file==1.5.1
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
robocorp-dialog==0.5.3
robotframework==5.0.1
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.4
robotframework-seleniumlibrary==6.0.0
robotframework-seleniumtestability==2.1.0
rpaframework==19.2.0
rpaframework-core==10.4.3
rpaframework-dialogs==4.0.3
rpaframework-pdf==5.0.1
rpaframework-windows==6.2.0
selenium==4.9.1
simple-salesforce==1.12.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stringcase==1.2.0
tenacity==8.2.2
termcolor==2.3.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==3.10.0
tzdata==2023.3
tzlocal==4.3
uiautomation==2.0.18
urllib3==2.0.2
webdriver-manager==3.8.6
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.15.0
